backend/requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==5.0.6
django-debug-toolbar==4.4.2
django-filter==24.2
djangorestframework==3.15.1
filelock==3.14.0
gunicorn==22.0.0
packaging==24.0
pbr==6.0.0
platformdirs==4.2.2
psycopg2-binary==2.9.9
sqlparse==0.5.0
stevedore==5.2.0
virtualenv==20.26.2
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.0
